refactor to remove handlers
