+++ src/main/kotlin/com/fcfb/arceus/controllers/UserController.kt
@@ -39,9 +39,9 @@ class UserController(
     fun getNewSignups() = userService.getNewSignups()
 
     @GetMapping("/name")
-    fun getUserByName(
+    fun getUserDTOByName(
         @RequestParam name: String,
-    ) = userService.getUserByName(name)
+    ) = userService.getUserDTOByName(name)
 
     @PutMapping("/update/password")
     fun updateUserPassword(
