+++ src/main/kotlin/com/fcfb/arceus/domain/Game.kt
@@ -585,6 +585,7 @@ class Game {
         CHEW("CHEW"),
         HURRY("HURRY"),
         NORMAL("NORMAL"),
+        NONE("NONE"),
     }
 
     enum class Scenario(val description: String) {
